poll-backend/requirements.txt
Django==2.0
django-cors-headers==2.1.0
django-debug-toolbar==1.9.1
djangorestframework==3.7.7
psycopg2==2.7.3.2
pytz==2017.3
sqlparse==0.2.4
